+++ home/noghartt/features/desktop/applications.nix
@@ -1,11 +1,9 @@
 { pkgs, ... }:
 {
   home.packages = with pkgs; [
-    calibre
     discord
     slack
     spotify
-    todoist-electron
     zotero
   ];
 }