+++ home/noghartt/features/desktop/default.nix
@@ -1,14 +1,10 @@
 {
   imports = [
-    ./hyprland.nix
-    ./noctalia.nix
-    ./rofi.nix
-    ./ghostty.nix
-    ./firefox.nix
-    ./chrome.nix
-    ./obsidian.nix
-    ./flameshot.nix
     ./applications.nix
-    ./mimeapps.nix
+    ./chrome.nix
+    ./firefox.nix
+    ./flameshot.nix
+    ./ghostty.nix
+    ./obsidian.nix
   ];
 }